// backend/src/db/schema.d.ts
export declare const sportLevelEnum: import("drizzle-orm/pg-core").PgEnum<["BEGINNER", "INTERMEDIATE", "ADVANCED", "EXPERT"]>;
// ...
export declare const sports: import("drizzle-orm/pg-core").PgTableWithColumns<{
    name: "sports";
    schema: undefined;
    columns: {
        id: import("drizzle-orm/pg-core").PgBuildColumn<"sports", import("drizzle-orm/pg-core").SetIsPrimaryKey<import("drizzle-orm/pg-core").SetHasDefault<import("drizzle-orm/pg-core").PgUUIDBuilder>>, {
            name: string;
            tableName: "sports";
            dataType: "string uuid";
            data: string;
            driverParam: string;
            notNull: true;
            hasDefault: true;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            identity: undefined;
            generated: undefined;
        }>;
        name: import("drizzle-orm/pg-core").PgBuildColumn<"sports", import("drizzle-orm/pg-core").SetNotNull<import("drizzle-orm/pg-core").PgVarcharBuilder<[string, ...string[]]>>, {
            name: string;
            tableName: "sports";
            dataType: "string";
            data: string;
            driverParam: string;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            identity: undefined;
            generated: undefined;
        }>;
        createdAt: import("drizzle-orm/pg-core").PgBuildColumn<"sports", import("drizzle-orm/pg-core").SetNotNull<import("drizzle-orm/pg-core").SetHasDefault<import("drizzle-orm/pg-core").PgTimestampBuilder>>, {
            name: string;
            tableName: "sports";
            dataType: "object date";
            data: Date;
            driverParam: string;
            notNull: true;
            hasDefault: true;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            identity: undefined;
            generated: undefined;
        }>;
    };
    dialect: "pg";
}>;
export declare const userSports: import("drizzle-orm/pg-core").PgTableWithColumns<{
    name: "user_sports";
    schema: undefined;
    columns: {
        userId: import("drizzle-orm/pg-core").PgBuildColumn<"user_sports", import("drizzle-orm/pg-core").SetNotNull<import("drizzle-orm/pg-core").PgUUIDBuilder>, {
            name: string;
            tableName: "user_sports";
            dataType: "string uuid";
            data: string;
            driverParam: string;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            identity: undefined;
            generated: undefined;
        }>;
        sportId: import("drizzle-orm/pg-core").PgBuildColumn<"user_sports", import("drizzle-orm/pg-core").SetNotNull<import("drizzle-orm/pg-core").PgUUIDBuilder>, {
            name: string;
            tableName: "user_sports";
            dataType: "string uuid";
            data: string;
            driverParam: string;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            identity: undefined;
            generated: undefined;
        }>;
        level: import("drizzle-orm/pg-core").PgBuildColumn<"user_sports", import("drizzle-orm/pg-core").SetNotNull<import("drizzle-orm/pg-core").PgEnumColumnBuilder<["BEGINNER", "INTERMEDIATE", "ADVANCED", "EXPERT"]>>, {
            name: string;
            tableName: "user_sports";
            dataType: "string enum";
            data: "BEGINNER" | "INTERMEDIATE" | "ADVANCED" | "EXPERT";
            driverParam: string;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: ["BEGINNER", "INTERMEDIATE", "ADVANCED", "EXPERT"];
            identity: undefined;
            generated: undefined;
        }>;
    };
    dialect: "pg";
}>;